Added function that downloads and installs tools.
Required some changes to URLs of Xcode command line tools binaries.
I may want to change the URLs again, now that I've better
acquainted myself with the curl command line options, and realize
that my problems with the Google Drive URLs were probably due to
not invoking curl with the -L and -k options.

This function is only partially implemented (the rest of its
implementation is commented out) and only partially tested
(because it requires downloading a large file). The test harness
itself is written with my personal laptop in mind; OS-dependent
values need to be changed depending on the machine used for testing.

I still need to test this function more completely.